refactor: 네비게이션 구조 개편
